requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
attrs=19.3.0=py_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.1=pypi_0
blas=1.0=mkl
bleach=3.1.5=py_0
brotlipy=0.7.0=py36h8c4c3a4_1000
bs4=0.0.1=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.7.22=0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=py36_0
cffi=1.14.0=py36h2e261b9_0
chardet=3.0.4=py36h9f0ad1d_1006
cloudpickle=1.4.1=py_0
cnn-finetune=0.6.0=pypi_0
cryptography=2.9.2=py36h45558ae_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py36_0
cytoolz=0.10.1=py36h7b6447c_0
dask-core=2.18.1=py_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
google-images-download=2.8.0=pypi_0
googleimagedownloader=4.0.0=pypi_0
graphite2=1.3.13=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.9=pypi_0
imageio=2.8.0=py_0
importlib-metadata=1.7.0=py36_0
importlib_metadata=1.7.0=0
intel-openmp=2019.4=243
ipykernel=5.3.2=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
jasper=2.0.14=h07fcdf6_1
jedi=0.17.1=py36_0
jinja2=2.11.2=py_0
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=6.1.6=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py36_0
kiwisolver=1.2.0=py36hfd86e86_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mistune=0.8.4=py36h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
munch=2.5.0=pypi_0
nbconvert=5.6.1=py36_0
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
networkx=2.4=py_0
ninja=1.9.0=py36hfd86e86_0
notebook=6.0.3=py36_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1h=h7b6447c_0
packaging=20.4=py_0
pandas=1.0.1=py36h0573a6f_0
pandas_ml=0.6.1=py36_0
pandoc=2.10=0
pandocfilters=1.4.2=py36_1
parso=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=6.2.0=py36h34e0f95_0
pip=20.1.1=py36_1
pixman=0.38.0=h7b6447c_0
pretrainedmodels=0.7.4=pypi_0
progressbar=2.5=py_0
progressbar2=3.51.0=pyh9f0ad1d_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.5=py_0
prompt_toolkit=3.0.5=0
ptyprocess=0.6.0=py36_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.16.0=py36h7b6447c_0
pysocks=1.7.1=py36h9f0ad1d_1
python=3.6.10=h0371630_0
python-dateutil=2.8.1=py_0
python-utils=2.3.0=py36_0
python_abi=3.6=1_cp36m
pytorch=1.3.1=cuda100py36h53c1284_0
pytorch-gpu=1.3.1=0
pytorch-metric-learning=0.9.88=pyh39e3cac_0
pytz=2020.1=py_0
pywavelets=1.1.1=py36h7b6447c_0
pyyaml=5.3.1=py36h7b6447c_0
pyzmq=19.0.1=py36h9947dbf_0
qt=5.9.7=h5867ecd_1
qtconsole=4.7.5=py_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
requests=2.23.0=pypi_0
requests-futures=1.0.0=pypi_0
scikit-image=0.16.2=py36h0573a6f_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.4.1=py36h0b6359f_0
seaborn=0.10.0=py_0
selenium=3.141.0=pypi_0
send2trash=1.5.0=py36_0
setuptools=47.3.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
soupsieve=2.0.1=pypi_0
sqlite=3.31.1=h62c20be_1
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchfile=0.1.0=py_0
torchvision=0.6.0=pypi_0
tornado=6.0.4=py36h7b6447c_1
tqdm=4.43.0=py_0
traitlets=4.3.3=py36_0
urllib3=1.25.9=py_0
visdom=0.1.8.9=0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
websocket-client=0.57.0=py36h9f0ad1d_1
wheel=0.34.2=py36_0
widgetsnbextension=3.5.1=py36_0
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
zeromq=4.3.2=he1b5a44_2
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
